# backend/boxen/assets/openapi.yaml
openapi: 3.1.2
...
info:
  title: Boxen Local API
  version: 1.0.0
  summary: Same-origin API for the locally hosted Boxen inventory system.
  description: >-
    This contract is normative for Boxen v1 HTTP field names and wire formats.
    The API is served only from the configured local Boxen origin: trusted-LAN
    HTTP is supported in development; production requires HTTPS.
    Anonymous access is configured server-side as off, viewer, or editor (default).
    When enabled, GET /session issues a per-browser session cookie and CSRF token
    without login. Other protected operations still require that session cookie.
    Anonymous access never grants user management, system administration, backup,
    or permanent purge permissions.
servers:
  - url: /api/v1
    description: Same-origin local Boxen server
...
security:
  - cookieSession: []
paths:
...
  /auth/oidc/{provider_id}/start:
    post:
      tags:
      - Authentication
      operationId: startOIDCSignIn
      summary: Begin browser-bound authorization-code sign-in with PKCE.
      security: []
      parameters:
      - name: provider_id
        in: path
        required: true
        schema:
          type: string
          pattern: ^[a-z][a-z0-9_-]{0,39}$
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: []
              properties: {}
      responses:
        '200':
          description: Operation completed.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                - authorization_url
                properties:
                  authorization_url:
                    type: string
  /auth/oidc/{provider_id}/callback:
    get:
      tags:
      - Authentication
      operationId: completeOIDCSignIn
      summary: Validate a one-time provider response and redirect to Boxen.
      security: []
      parameters:
      - name: provider_id
        in: path
        required: true
        schema:
          type: string
      - name: code
        in: query
        required: false
        schema:
          type: string
      - name: state
        in: query
        required: false
        schema:
          type: string
      - name: error
        in: query
        required: false
        schema:
          type: string
      responses:
        '302':
          description: Operation completed.
  /auth/provider-status:
    get:
      tags:
      - Authentication
      operationId: listAuthProviderStatus
      summary: Inspect configured sign-in providers (owner only).
      responses:
        '200':
          description: Operation completed.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuthProviderStatusView'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /auth/identities:
    get:
      tags:
      - Authentication
      operationId: listAuthIdentities
      summary: List explicit provider subject bindings (owner only).
      parameters:
      - name: user_id
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/Uuid'
      responses:
        '200':
          description: Operation completed.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required:
                - items
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuthIdentityView'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags:
      - Authentication
      operationId: linkAuthIdentity
      summary: Bind a provider subject to an existing local user after recent owner
        sign-in.
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
              - user_id
              - provider_id
              - subject
              properties:
                user_id:
                  $ref: '#/components/schemas/Uuid'
                provider_id:
                  type: string
                subject:
                  type: string
                  minLength: 1
                  maxLength: 255
      responses:
        '201':
          description: Operation completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthIdentityView'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /auth/identities/{identity_id}:
    delete:
      tags:
      - Authentication
      operationId: unlinkAuthIdentity
      summary: Unlink an identity and revoke the user sessions after recent owner sign-in.
      parameters:
      - name: identity_id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
      - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204':
          description: Operation completed.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: boxen_session
      description: >-
        HttpOnly, SameSite=Strict session cookie, Secure over HTTPS. Issued by
        local login, first-owner setup, or GET /session when anonymous access is
        enabled. Anonymous mutations retain the same Origin, CSRF, role, and
        resource-version checks as local user sessions.
...
  parameters:
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string, minLength: 32, maxLength: 256 }
...
  responses:
...
    Unauthorized:
      description: Authentication is required or no longer valid.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Forbidden:
      description: The current user is not authorized.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
...
  schemas:
    Uuid:
      type: string
      format: uuid
    DateTime:
      type: string
      format: date-time
...
    FieldError:
      type: object
      additionalProperties: false
      required: [path, code, message]
      properties:
        path: { type: string }
        code: { type: string }
        message: { type: string }
    ProblemDetails:
      type: object
      additionalProperties: true
      required: [type, title, status, code, request_id]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        code: { type: string }
        request_id: { type: string }
        errors:
          type: array
          items: { $ref: '#/components/schemas/FieldError' }
...
    AuthProviderStatusView:
      type: object
      additionalProperties: false
      required:
      - id
      - label
      - issuer
      - client_id
      - configured
      - redirect_uri
      properties:
        id:
          type: string
        label:
          type: string
        issuer:
          type: string
        client_id:
          type: string
        configured:
          type: boolean
        redirect_uri:
          type: string
    AuthIdentityView:
      type: object
      additionalProperties: false
      required:
      - id
      - user_id
      - provider_id
      - issuer
      - subject
      - created_at
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        user_id:
          $ref: '#/components/schemas/Uuid'
        provider_id:
          type: string
        issuer:
          type: string
        subject:
          type: string
        created_at:
          $ref: '#/components/schemas/DateTime'
...